[CuTe] Mark existing Xe atoms as legacy
