+++ include/cute/atom/copy_atom.hpp
@@ -1,5 +1,6 @@
 /***************************************************************************************************
  * Copyright (c) 2023 - 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * Copyright (C) 2025 Intel Corporation, All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  *
  * Redistribution and use in source and binary forms, with or without
@@ -715,6 +716,7 @@ print(ThrCopy<TiledCopy, ThrIdx> const& thr_copy)
 
 #if defined(SYCL_INTEL_TARGET)
 #include <cute/atom/copy_traits_xe.hpp>
+#include <cute/atom/copy_traits_xe_legacy.hpp>
 #endif
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
